requirements.txt
PyQt5==5.15.10
requests==2.31.0
psutil==5.9.8
pyinstaller==6.19.0
Pillow>=10.0.0
